Moves console output out of linter
