+++ lib/config.js
@@ -17,6 +17,7 @@ class Config {
             lint: {
                 rules: this.notEmptyArray(args.rules),
                 skip: this.notEmptyArray(args.skip),
+                output: args.output
             },
             resolve: {
                 output: args.output,
