+++ lint.js
@@ -8,69 +8,19 @@ const linter = require('oas-linter');
 const rules = require('./lib/rules.js');
 const validator = require('oas-validator');
 const fromJsonSchema = require('json-schema-to-openapi-schema');
+const consoleOutputRenderer = require('./lib/output/console.js');
 
-const colors = process.env.NODE_DISABLE_COLORS ? {} : {
-    red: '\x1b[31m',
-    green: '\x1b[32m',
-    yellow: '\x1b[33m',
-    blue: '\x1b[34m',
-    magenta: '\x1b[35m',
-    cyan: '\x1b[36m',
-    white: '\x1b[37m',
-    reset: '\x1b[0m'
-};
-
-const formatSchemaError = (err, context) => {
-    const pointer = context.pop();
-    let output = `
-${colors.yellow + pointer}
-`;
-    
-    if (err.name === 'AssertionError' || err.error.name === 'AssertionError') {
-        output += colors.reset + truncateLongMessages(err.message);
-    }
-    else if (err instanceof validator.CLIError) {
-        output += colors.reset + err.message;
-    }
-    else {
-        output += colors.red + err.stack;
-    }
-    return output;
-}
-
-const truncateLongMessages = message => {
-    let lines = message.split('\n');
-    if (lines.length > 6) {
-        lines = lines.slice(0, 5).concat(
-            ['  ... snip ...'],
-            lines.slice(-1)
-        );
-    }
-    return lines.join('\n');
-}
-
-const formatLintResults = lintResults => {
-    let output = '';
-    lintResults.forEach(result => {
-        const { rule, error, pointer } = result;
-
-        output += `
-${colors.yellow + pointer} ${colors.cyan} R: ${rule.name} ${colors.white} D: ${rule.description}
-${colors.reset + truncateLongMessages(error.message)}
-
-More information: ${rule.url}#${rule.name}
-`;
-    });
-
-    return output;
+const getOutputRenderer = type => {
+    return consoleOutputRenderer;
 }
 
 const command = async (specFile, cmd) => {
-    config.init(cmd);
+    config.init(cmd);    
     const jsonSchema = config.get('jsonSchema');
     const verbose = config.get('quiet') ? 0 : config.get('verbose', 1);
     const rulesets = config.get('lint:rules', []);
     const skip = config.get('lint:skip', []);
+    const outputRenderer = getOutputRenderer(config.get('lint:output'));
 
     rules.init({
         skip
@@ -87,26 +37,14 @@ const command = async (specFile, cmd) => {
         validator.validate(spec, buildValidatorOptions(skip, verbose), (err, _options) => {
             const { context, warnings, valid } = _options || err.options;
 
+            outputRenderer.render(err, warnings, valid, context, cmd.quiet);
+            
             if (err && valid === false) {
-                console.error(colors.red + 'Specification schema is invalid.' + colors.reset);
-                if (err.name === 'AssertionError') {
-                    console.error(formatSchemaError(err, context));
-                }
-
-                for (let linterResult of err.options.linterResults()) {
-                    console.error(formatSchemaError(linterResult, context));
-                }
                 return reject();
             }
 
             if (warnings.length) {
-                console.error(colors.red + 'Specification contains lint errors: ' + warnings.length + colors.reset);
-                console.warn(formatLintResults(warnings))
                 return reject();
-            }
-
-            if (!cmd.quiet) {
-                console.log(colors.green + 'Specification is valid, with 0 lint errors' + colors.reset)
             }
 
             return resolve();
